+++ internal/domain/user.go
@@ -3,5 +3,5 @@ package domain
 type UserJwt struct {
 	Account string `json:"account"`
 	Role    int64  `json:"role"`
-	Id      uint   `json:"id"`
+	Id      int64  `json:"id"`
 }
